move database functions to tools
