--- packages/base/src/dialogs/symbology/utils/indexedDBUtil.ts
@@ -1,43 +1,0 @@
-export const openDatabase = () => {
-  return new Promise<IDBDatabase>((resolve, reject) => {
-    const request = indexedDB.open('GeoTIFFCache', 1);
-
-    request.onupgradeneeded = event => {
-      const db = request.result;
-      if (!db.objectStoreNames.contains('files')) {
-        db.createObjectStore('files', { keyPath: 'url' });
-      }
-    };
-
-    request.onsuccess = () => resolve(request.result);
-    request.onerror = () => reject(request.error);
-  });
-};
-
-export const saveToIndexedDB = async (
-  key: string,
-  file: Blob,
-  metadata: any
-) => {
-  const db = await openDatabase();
-  return new Promise<void>((resolve, reject) => {
-    const transaction = db.transaction('files', 'readwrite');
-    const store = transaction.objectStore('files');
-    store.put({ url: key, file, metadata });
-
-    transaction.oncomplete = () => resolve();
-    transaction.onerror = () => reject(transaction.error);
-  });
-};
-
-export const getFromIndexedDB = async (key: string) => {
-  const db = await openDatabase();
-  return new Promise<any>((resolve, reject) => {
-    const transaction = db.transaction('files', 'readonly');
-    const store = transaction.objectStore('files');
-    const request = store.get(key);
-
-    request.onsuccess = () => resolve(request.result);
-    request.onerror = () => reject(request.error);
-  });
-};
